Move query to experimental
