+++ ql/test/experimental/CWE-681/IncorrectNumericConversion.qlref
@@ -1,0 +1,1 @@
+experimental/CWE-681/IncorrectNumericConversion.ql
