--- ql/test/query-tests/Security/CWE-681/IncorrectNumericConversion.qlref
@@ -1,1 +1,0 @@
-Security/CWE-681/IncorrectNumericConversion.ql
